+++ hydromt/data_catalog/drivers/dataset/dataset_driver.py
@@ -28,7 +28,6 @@ class DatasetDriver(BaseDriver, ABC):
         variables: Optional[Variables] = None,
         time_range: Optional[TimeRange] = None,
         metadata: Optional[SourceMetadata] = None,
-        logger: Logger = logger,
         handle_nodata: NoDataStrategy = NoDataStrategy.RAISE,
         # TODO: https://github.com/Deltares/hydromt/issues/802
     ) -> xr.Dataset:
@@ -37,7 +36,7 @@ class DatasetDriver(BaseDriver, ABC):
 
         args:
         """
-        uris = self.metadata_resolver.resolve(
+        uris = self.uri_resolver.resolve(
             uri,
             self.filesystem,
             time_range=time_range,
@@ -50,7 +49,6 @@ class DatasetDriver(BaseDriver, ABC):
             time_range=time_range,
             variables=variables,
             metadata=metadata,
-            logger=logger,
             handle_nodata=handle_nodata,
         )
 
@@ -62,7 +60,6 @@ class DatasetDriver(BaseDriver, ABC):
         variables: Optional[Variables] = None,
         time_range: Optional[TimeRange] = None,
         metadata: Optional[SourceMetadata] = None,
-        logger: Logger,
         handle_nodata: NoDataStrategy = NoDataStrategy.RAISE,
     ) -> xr.Dataset:
         """