+++ hydromt/model/model.py
@@ -42,7 +42,7 @@ __all__ = ["Model"]
 # see also hydromt.model group in pyproject.toml
 __hydromt_eps__ = ["Model"]
 
-_logger = logging.getLogger(__name__)
+logger = logging.getLogger(__name__)
 T = TypeVar("T", bound=ModelComponent)
 
 
@@ -62,7 +62,6 @@ class Model(object, metaclass=ABCMeta):
         mode: str = "w",
         data_libs: Optional[Union[List, str]] = None,
         region_component: Optional[str] = None,
-        logger=_logger,
         **catalog_keys,
     ):
         """Initialize a model.
@@ -105,12 +104,8 @@ class Model(object, metaclass=ABCMeta):
 
         data_libs = data_libs or []
 
-        self.logger = logger
-
         # link to data
-        self.data_catalog = DataCatalog(
-            data_libs=data_libs, logger=self.logger, **catalog_keys
-        )
+        self.data_catalog = DataCatalog(data_libs=data_libs, **catalog_keys)
         """DataCatalog for data access"""
 
         # file system
@@ -123,7 +118,7 @@ class Model(object, metaclass=ABCMeta):
         self._defered_file_closes: List[DeferedFileClose] = []
 
         model_metadata = cast(Dict[str, str], PLUGINS.model_metadata[self.name])
-        self.logger.info(
+        logger.info(
             f"Initializing {self.name} model from {model_metadata['plugin_name']} (v{model_metadata['version']})."
         )
 
@@ -152,7 +147,7 @@ class Model(object, metaclass=ABCMeta):
                     "Specify region_component."
                 )
             if len(has_region_components) == 0:
-                self.logger.warning("No region component found in components.")
+                logger.warning("No region component found in components.")
                 return ""
             return has_region_components[0][0]
 
@@ -266,7 +261,7 @@ class Model(object, metaclass=ABCMeta):
 
         for step_dict in steps:
             step, kwargs = next(iter(step_dict.items()))
-            self.logger.info(f"build: {step}")
+            logger.info(f"build: {step}")
             # Call the methods.
             method = _rgetattr(self, step)
             params = {
@@ -276,7 +271,7 @@ class Model(object, metaclass=ABCMeta):
             }
             merged = {**params, **kwargs}
             for k, v in merged.items():
-                self.logger.info(f"{method}.{k}: {v}")
+                logger.info(f"{method}.{k}: {v}")
             method(**kwargs)
 
         # If there are any write options included in the steps,
@@ -352,7 +347,7 @@ class Model(object, metaclass=ABCMeta):
         # loop over methods from config file
         for step_dict in steps:
             step, kwargs = next(iter(step_dict.items()))
-            self.logger.info(f"update: {step}")
+            logger.info(f"update: {step}")
             # Call the methods.
             method = _rgetattr(self, step)
             params = {
@@ -362,7 +357,7 @@ class Model(object, metaclass=ABCMeta):
             }
             merged = {**params, **kwargs}
             for k, v in merged.items():
-                self.logger.info(f"{method}.{k}: {v}")
+                logger.info(f"{method}.{k}: {v}")
             method(**kwargs)
 
         # If there are any write options included in the steps,
@@ -398,7 +393,7 @@ class Model(object, metaclass=ABCMeta):
                 the components that should be read from disk. If None is provided
                 all components will be read.
         """
-        self.logger.info(f"Reading model data from {self.root.path}")
+        logger.info(f"Reading model data from {self.root.path}")
         components = components or list(self.components.keys())
         for c in [self.components[name] for name in components]:
             c.read()
@@ -437,7 +432,7 @@ class Model(object, metaclass=ABCMeta):
         """
         self.root._assert_write_mode()
         path = data_lib_fn if isabs(data_lib_fn) else join(self.root.path, data_lib_fn)
-        cat = DataCatalog(logger=self.logger, fallback_lib=None)
+        cat = DataCatalog(fallback_lib=None)
         # read hydromt_data yml file and add to data catalog
         if self.root.is_reading_mode() and isfile(path) and append:
             cat.from_yml(path)