+++ hydromt/model/processes/basin_mask.py
@@ -32,7 +32,6 @@ def get_basin_geometry(
     basins_name="basins",
     flwdir_name="flwdir",
     ftype="infer",
-    logger=logger,
     buffer=10,
     **stream_kwargs,
 ):
@@ -147,7 +146,7 @@ def get_basin_geometry(
                     "Basin geometries ignored as no corresponding"
                     + " basin map is provided."
                 )
-            _check_size(ds, logger)
+            _check_size(ds)
             logger.info(f'basin map "{basins_name}" missing, calculating on the fly.')
             flwdir = flwdir_from_da(ds[flwdir_name], ftype=ftype)
             ds[basins_name] = xr.Variable(ds.raster.dims, flwdir.basins())
@@ -219,7 +218,7 @@ def get_basin_geometry(
     xy_out = None
     if kind in ["subbasin", "interbasin"]:
         # get flow directions
-        _check_size(ds, logger)  # warning for large domain
+        _check_size(ds)  # warning for large domain
         mask = False
         # if interbasin, set flwdir mask within geometry / bounding box
         if kind == "interbasin":  # we have checked before that geom is not None
@@ -285,7 +284,7 @@ def get_basin_geometry(
     return basin_geom, outlet_geom
 
 
-def _check_size(ds, logger=logger, threshold=12e3**2):
+def _check_size(ds, threshold=12e3**2):
     # warning for large domain
     if (
         np.multiply(*ds.raster.shape) > threshold