+++ hydromt/model/processes/region.py
@@ -23,7 +23,7 @@ from hydromt.plugins import PLUGINS
 if TYPE_CHECKING:
     from hydromt.model.model import Model
 
-_logger = getLogger(__name__)
+logger: Logger = getLogger(__name__)
 
 __all__ = [
     "parse_region_basin",
@@ -41,7 +41,6 @@ def parse_region_basin(
     data_catalog: DataCatalog,
     hydrography_fn: StrPath,
     basin_index_fn: Optional[StrPath] = None,
-    logger: Logger = _logger,
 ) -> gpd.GeoDataFrame:
     """Parse a basin /subbasin / interbasin region and return the GeoDataFrame.
 
@@ -110,8 +109,6 @@ def parse_region_basin(
         Path of the hydrography raster dataset in the data catalog.
     basin_index_fn : strPath, optional
         Path of the basin index raster dataset in the data catalog.
-    logger : Logger, optional
-        Logger object.
     """
     var_thresh_kwargs = region.copy()
     kind = next(iter(region))
@@ -143,17 +140,12 @@ def parse_region_basin(
     if "bounds" not in kwargs and basin_index_fn is not None:
         kwargs.update(basin_index=data_catalog.get_source(str(basin_index_fn)))
     # get basin geometry
-    geom, _ = get_basin_geometry(ds=ds_org, kind=kind, logger=logger, **kwargs)
+    geom, _ = get_basin_geometry(ds=ds_org, kind=kind, **kwargs)
 
     return geom
 
 
-def parse_region_bbox(
-    region: dict,
-    *,
-    crs: int = 4326,
-    logger: Logger = _logger,
-) -> gpd.GeoDataFrame:
+def parse_region_bbox(region: dict, *, crs: int = 4326) -> gpd.GeoDataFrame:
     """Parse a region of kind bbox and return the GeoDataFrame.
 
     Parameters
@@ -164,8 +156,6 @@ def parse_region_bbox(
     * {'bbox': [xmin, ymin, xmax, ymax]}
     crs : CRS, optional
         CRS of the bounding box coordinates. By default EPSG 4326.
-    logger : Logger, optional
-        Logger object.
     """
     kwargs = region.copy()
     kind = next(iter(region))
@@ -191,7 +181,6 @@ def parse_region_geom(
     *,
     crs: Optional[int] = None,
     data_catalog: Optional[DataCatalog] = None,
-    logger: Logger = _logger,
 ) -> gpd.GeoDataFrame:
     """Parse a region and return the GeoDataFrame.
 
@@ -205,8 +194,6 @@ def parse_region_geom(
         CRS of the geometry in case it cannot be derived from the geometry itself.
     data_catalog : DataCatalog, optional
         DataCatalog object containing the data sources.
-    logger : Logger, optional
-        Logger object.
     """
     kwargs = region.copy()
     kind = next(iter(region))
@@ -239,7 +226,6 @@ def parse_region_grid(
     region: dict,
     *,
     data_catalog: Optional[DataCatalog],
-    logger: Logger = _logger,
 ) -> Union[xr.DataArray, xr.Dataset]:
     """Parse a region of kind grid and return the corresponding xarray object.
 
@@ -256,8 +242,6 @@ def parse_region_grid(
         * {'grid': /path/to/grid}
     data_catalog : DataCatalog
         DataCatalog object containing the data sources.
-    logger : Logger, optional
-        Logger object.
 
     Returns
     -------
@@ -288,7 +272,7 @@ def parse_region_grid(
     return da
 
 
-def parse_region_other_model(region: dict, *, logger: Logger = _logger) -> "Model":
+def parse_region_other_model(region: dict) -> "Model":
     """Parse a region with a model path and return that whole Model in read mode.
 
     Parameters
@@ -297,8 +281,6 @@ def parse_region_other_model(region: dict, *, logger: Logger = _logger) -> "Mode
         Dictionary describing region of interest.
         For a region based of another models grid:
         * {'<model_name>': root}
-    logger : Logger, optional
-        Logger object.
     """
     kwargs = region.copy()
     kind = next(iter(region))
@@ -308,7 +290,7 @@ def parse_region_other_model(region: dict, *, logger: Logger = _logger) -> "Mode
     logger.debug(f"Parsed region (kind={kind}): {str(kwargs)}")
 
     model_class = PLUGINS.model_plugins[kind]
-    return model_class(root=value0, mode="r", logger=logger)
+    return model_class(root=value0, mode="r")
 
 
 def parse_region_mesh(region: dict) -> xu.UgridDataset: