+++ tests/conftest.py
@@ -1,4 +1,3 @@
-import logging
 from os import sep
 from os.path import abspath, dirname, join
 from pathlib import Path
@@ -28,7 +27,7 @@ from hydromt.data_catalog.adapters.geodataframe import GeoDataFrameAdapter
 from hydromt.data_catalog.adapters.geodataset import GeoDatasetAdapter
 from hydromt.data_catalog.drivers import GeoDataFrameDriver, RasterDatasetDriver
 from hydromt.data_catalog.drivers.geodataset.geodataset_driver import GeoDatasetDriver
-from hydromt.data_catalog.uri_resolvers import MetaDataResolver
+from hydromt.data_catalog.uri_resolvers import URIResolver
 from hydromt.gis.raster_utils import affine_to_coords
 from hydromt.model.components.config import ConfigComponent
 from hydromt.model.components.geoms import GeomsComponent
@@ -506,14 +505,14 @@ def vector_model_no_defaults(ts, geodf, mocker: MockerFixture):
 
 
 @pytest.fixture()
-def mock_resolver() -> MetaDataResolver:
-    class MockMetaDataResolver(MetaDataResolver):
+def mock_resolver() -> URIResolver:
+    class MockURIResolver(URIResolver):
         name = "mock_resolver"
 
         def resolve(self, uri, *args, **kwargs):
             return [uri]
 
-    resolver = MockMetaDataResolver()
+    resolver = MockURIResolver()
     return resolver
 
 
@@ -539,7 +538,7 @@ def mock_geo_ds_adapter():
 
 @pytest.fixture()
 def mock_geodf_driver(
-    geodf: gpd.GeoDataFrame, mock_resolver: MetaDataResolver
+    geodf: gpd.GeoDataFrame, mock_resolver: URIResolver
 ) -> GeoDataFrameDriver:
     class MockGeoDataFrameDriver(GeoDataFrameDriver):
         name = "mock_geodf_driver"
@@ -547,12 +546,12 @@ def mock_geodf_driver(
         def read_data(self, *args, **kwargs) -> gpd.GeoDataFrame:
             return geodf
 
-    return MockGeoDataFrameDriver(metadata_resolver=mock_resolver)
+    return MockGeoDataFrameDriver(uri_resolver=mock_resolver)
 
 
 @pytest.fixture()
 def mock_raster_ds_driver(
-    raster_ds: xr.Dataset, mock_resolver: MetaDataResolver
+    raster_ds: xr.Dataset, mock_resolver: URIResolver
 ) -> RasterDatasetDriver:
     class MockRasterDatasetDriver(RasterDatasetDriver):
         name = "mock_raster_ds_driver"
@@ -561,12 +560,12 @@ def mock_raster_ds_driver(
         def read_data(self, *args, **kwargs) -> xr.Dataset:
             return raster_ds
 
-    return MockRasterDatasetDriver(metadata_resolver=mock_resolver)
+    return MockRasterDatasetDriver(uri_resolver=mock_resolver)
 
 
 @pytest.fixture()
 def mock_geo_ds_driver(
-    geoda: xr.DataArray, mock_resolver: MetaDataResolver
+    geoda: xr.DataArray, mock_resolver: URIResolver
 ) -> GeoDatasetDriver:
     class MockGeoDatasetDriver(GeoDatasetDriver):
         name = "mock_geo_ds_driver"
@@ -575,7 +574,7 @@ def mock_geo_ds_driver(
         def read_data(self, *args, **kwargs) -> xr.Dataset:
             return geoda.to_dataset()
 
-    return MockGeoDatasetDriver(metadata_resolver=mock_resolver)
+    return MockGeoDatasetDriver(uri_resolver=mock_resolver)
 
 
 @pytest.fixture()
@@ -587,13 +586,10 @@ def artifact_data():
 
 @pytest.fixture()
 def mock_model(tmpdir, mocker: MockerFixture):
-    logger = logging.getLogger(__name__)
-    logger.propagate = True
     model = mocker.create_autospec(Model)
     model.root = mocker.create_autospec(ModelRoot(tmpdir), instance=True)
     model.root.path.return_value = tmpdir
     model.data_catalog = mocker.create_autospec(DataCatalog)
-    model.logger = logger
     return model